Remove config option for command aliases
This commit removes the config options `commands.alias.ecs-shop` and `commands.alias.ecsadmin-adminshop`, and adds warnings for servers that use either option.

If you prefer these command aliases, you can edit the `commands.yml` file in your server root directory to configure command aliases.

The motivation for this change is deprecation of the `SimplePluginManager` class in Paper that was being reflected into in order to add aliases.